Rename shm to ipc to better reflect purpose
