+++ pymapd/ipc.py
